+++ scripts/fobi_sync_plugins.sh
@@ -1,0 +1,2 @@
+cd examples/simple/
+./manage.py fobi_sync_plugins --traceback -v 3