+++ src/fobi/contrib/plugins/form_elements/fields/duration/__init__.py
@@ -1,0 +1,10 @@
+__title__ = 'fobi.contrib.plugins.form_elements.fields.duration'
+__author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
+__copyright__ = '2014-2017 Artur Barseghyan'
+__license__ = 'GPL 2.0/LGPL 2.1'
+__all__ = ('default_app_config', 'UID',)
+
+default_app_config = 'fobi.contrib.plugins.form_elements.fields.' \
+                     'duration.apps.Config'
+
+UID = 'duration'